# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CheckoutRequest:
    properties:
      address_id:
        type: string
      delivery_method:
        enum:
        - instant
        - next_day
        - regular
        type: string
      discount_code:
        maxLength: 50
        type: string
    required:
    - address_id
    - delivery_method
    type: object
  dto.CreateProductRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      image_url:
        maxLength: 500
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      price:
        minimum: 0
        type: integer
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    type: object
...
  dto.PreviewRequest:
    properties:
      delivery_method:
        enum:
        - instant
        - next_day
        - regular
        type: string
      discount_code:
        maxLength: 50
        type: string
    required:
    - delivery_method
    type: object
...
  dto.UpdateProductRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      image_url:
        maxLength: 500
        type: string
      name:
        maxLength: 200
        minLength: 1
        type: string
      price:
        minimum: 0
        type: integer
      stock:
        minimum: 0
        type: integer
    required:
    - name
    - price
    type: object
...
  response.ErrorCode:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - CONFLICT
    - UNPROCESSABLE_ENTITY
    - INTERNAL_SERVER_ERROR
    - VALIDATION_ERROR
    - TOKEN_EXPIRED
    - TOKEN_INVALID
    type: string
    x-enum-varnames:
    - ErrCodeBadRequest
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeNotFound
    - ErrCodeConflict
    - ErrCodeUnprocessable
    - ErrCodeInternalServerError
    - ErrCodeValidation
    - ErrCodeTokenExpired
    - ErrCodeTokenInvalid
  response.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      error_code:
        $ref: '#/definitions/response.ErrorCode'
      list_data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      status:
        type: boolean
    type: object
info:
  contact:
    name: SEAPEDIA Dev
  description: |-
    Backend API marketplace multi-role (Admin, Seller, Buyer, Driver) — COMPFEST 18.
    Versi yang tampil di UI diambil dari config.Version (di-set saat runtime).
    Login pakai email. Otorisasi mengikuti ACTIVE ROLE pada JWT (pilih via /auth/select-role).
  termsOfService: http://swagger.io/terms/
  title: SEAPEDIA API
  version: "1.0"
paths:
...
  /buyer/checkout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Alamat + metode kirim + diskon opsional
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CheckoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Checkout dari cart (buyer)
      tags:
      - Buyer
  /buyer/checkout/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: Metode kirim + diskon opsional
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PreviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hitung ringkasan checkout (buyer)
      tags:
      - Buyer
  /buyer/orders:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Riwayat order (buyer)
      tags:
      - Buyer
  /buyer/orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail order + timeline status (buyer)
      tags:
      - Buyer
  /buyer/reports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ringkasan pengeluaran (buyer)
      tags:
      - Buyer
...
  /products:
    get:
      parameters:
      - description: Cari nama produk
        in: query
        name: search
        type: string
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Katalog produk (publik)
      tags:
      - Public
  /products/{id}:
    get:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Detail produk (publik)
      tags:
      - Public
...
  /seller/orders:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Daftar order masuk (seller)
      tags:
      - Seller
  /seller/orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail order toko (seller)
      tags:
      - Seller
  /seller/orders/{id}/process:
    post:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 'Proses order: Sedang Dikemas → Menunggu Pengirim (seller)'
      tags:
      - Seller
  /seller/products:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Daftar produk toko sendiri (seller)
      tags:
      - Seller
    post:
      consumes:
      - application/json
      parameters:
      - description: Produk
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tambah produk (seller)
      tags:
      - Seller
  /seller/products/{id}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus produk sendiri (seller)
      tags:
      - Seller
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Produk
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ubah produk sendiri (seller)
      tags:
      - Seller
  /seller/reports:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ringkasan pendapatan (seller)
      tags:
      - Seller
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"